refactor: Move dao setup code out of the Statebox class.
