+++ lib/index.js
@@ -9,16 +9,14 @@
 const executioner = require('./executioner')
 const StateMachines = require('./state-machines')
 const Resources = require('./resources')
-const MemoryDao = require('./dao/Memory-dao')
-const StorageDao = require('./dao/StorageService-dao')
 const Status = require('./Status')
 const CallbackManager = require('./Callback-manager')
-const debug = require('debug')('statebox')
+const findDao = require('./dao-loader')
 
 class Statebox {
   constructor (options = {}) {
     this.options = options
-    this.ready_ = this._findDao(options)
+    this.ready_ = findDao(options)
       .then(dao => {
         info(options.messages, 'Statebox is ready')
         this.options.dao = dao
@@ -30,47 +28,6 @@ class Statebox {
   get ready () {
     return this.ready_
   } // ready
-
-  async _findDao (options) {
-    if (options.dao) {
-      debug('Using custom Dao')
-      return options.dao // custom DAO provided
-    }
-
-    const storageDao = await this._daoFromStorage(options)
-    if (storageDao) {
-      debug('Using Storage Dao')
-      return storageDao
-    }
-
-    debug('Using built in MemoryDao')
-    return new MemoryDao(options)
-  } // _findDao
-
-  async _daoFromStorage (options) {
-    if (!options.bootedServices || !options.bootedServices.storage) {
-      return null
-    }
-    try {
-      const storage = options.bootedServices.storage
-      info(options.messages, `Using storage service ${storage.storageName}`)
-      let model = storage.models[StorageDao.ExecutionModelName]
-      if (!model) {
-        info(options.messages, `Adding model ${StorageDao.ExecutionModelName}`)
-        model = await storage.addModel(
-          StorageDao.ExecutionModelName,
-          StorageDao.ExecutionModelDefinition,
-          options.messages
-        )
-      }
-      return new StorageDao(model)
-    } catch (err) {
-      warning(options.messages, 'Could not get Dao from storage service')
-      warning(options.messages, err)
-      warning(options.messages, 'Falling back to in-memory Dao')
-    }
-    return null
-  } // _daoFromStorage
 
   createModuleResource (moduleName, moduleClass) {
     Resources.createModule(moduleName, moduleClass)
@@ -253,13 +210,5 @@ function info (messages, msg) {
     console.log(msg)
   }
 } // info
-function warning (messages, msg) {
-  if (messages) {
-    info(messages, msg)
-    messages.warning(msg)
-  } else {
-    console.log(msg)
-  }
-} // warning
 
 module.exports = Statebox
